+++ changelog.md
@@ -2,11 +2,13 @@
 ## [0.20.0] - UNRELEASED
 ### Breaking Changes
 * `FixedTimeLoadOverride` now takes in optional list of floats instead of optional float for its variable.
+* TODO: new generatorConfig fields.
 
 ### New Features
 * Modification to `ModelConfig` to allow more customization for model generation
   * Added support for separate `vMinPu` and `vMaxPu` with load and generators.
   * Added support for `ctPrimScalingFactor` which is required when calculating new `ctPrim` value when feeder proxy loads are not in use.
+* Support node level results
 
 ### Enhancements
 * Supports passing a list of feeders to `run_hosting_capacity_calibration` to perform calibration on a subset of the network.